% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data_transformed.R
\name{fw_auth}
\alias{fw_auth}
\title{Authenticate User and Set GITHUB_TOKEN}
\usage{
fw_auth()
}
\value{
Invisibly returns `TRUE` after successful authentication and token setup.
}
\description{
This function checks if a valid GITHUB_TOKEN is already set. If not, it prompts the user
to enter a password, which is used to generate and store the GITHUB_TOKEN for future access.
The GITHUB_TOKEN is required for accessing private repositories.
}
