% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data_transformed.R
\name{fw_data}
\alias{fw_data}
\title{Download CESAB Functionalwebs Data}
\usage{
fw_data(version = NULL, path = NULL, biomass = FALSE)
}
\arguments{
\item{version}{Version number. Default is `NULL` to load the most recent version.}

\item{path}{Path to store the data at. Defaults to `NULL`, in which case `datastorr` will use `rappdirs` for best storage location.}

\item{biomass}{Logical indicating if the biomass dataset should be accessed. Defaults to FALSE.}
}
\value{
The specified version of the dataset.
}
\description{
This function downloads a specific version of the CESAB Functionalwebs dataset.
If no version is provided, it will load the most recent version from either local storage or GitHub.
The `fw_del` function can be used to delete downloaded datasets.
}
