% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data_transformed.R
\name{fw_data_transformed}
\alias{fw_data_transformed}
\title{Load and Transform Data}
\usage{
fw_data_transformed(
  version = "0.7.7",
  path = NULL,
  biomass = FALSE,
  private = FALSE
)
}
\arguments{
\item{version}{Version of the dataset to load. Default is "0.7.7".}

\item{path}{Path to store the data at. Defaults to `NULL`, in which case the system's default path is used.}

\item{biomass}{Logical indicating if the biomass dataset should be accessed. Defaults to FALSE.}

\item{private}{Logical indicating if the dataset should be accessed privately. Defaults to FALSE.}
}
\value{
The transformed dataset, including additional trait corrections and transformations.
}
\description{
This function loads the CESAB Functionalwebs data and applies various transformations.
Additional files such as "Extra_traits.csv" and "taxon_level_traits.csv" are read and merged into the dataset.
It also applies specific data transformations based on traits, correcting certain values as specified by the user.
}
