% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data_transformed.R
\name{fw_info}
\alias{fw_info}
\title{Retrieve Functional Web Data Information}
\usage{
fw_info(path, biomass = FALSE)
}
\arguments{
\item{path}{Optional path to save the data.}

\item{biomass}{Logical indicating if the biomass dataset should be accessed. Defaults to FALSE.}
}
\value{
Returns the release information for the relevant dataset from GitHub.
}
\description{
This function fetches the information of a dataset from GitHub, specifically
the dataset related to either biomass or functional webs, depending on the provided argument.
}
