% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data_transformed.R
\name{fw_versions}
\alias{fw_versions}
\title{Get Available Versions of Data}
\usage{
fw_versions(local = TRUE, path = NULL, biomass = FALSE)
}
\arguments{
\item{local}{Logical indicating if local or GitHub versions should be polled. Defaults to TRUE.}

\item{path}{Optional path to save the data.}

\item{biomass}{Logical indicating if the biomass dataset should be accessed. Defaults to FALSE.}
}
\value{
A list of available versions of the data.
}
\description{
This function retrieves the available versions of data from either local storage
or GitHub, based on the specified parameters. If `local = FALSE`, the function will check for
the most recent version on GitHub.
}
