% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data_transformed.R
\name{fw_auth}
\alias{fw_auth}
\title{Asks for your password}
\usage{
fw_auth()
}
\description{
This password is only for members of the CESAB functional webs group. Please do not add it to your code!
}
