% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data.R
\name{fw_data}
\alias{fw_data}
\title{Download CESAB Functionalwebs Data}
\usage{
fw_data(version = NULL, path = NULL, biomass = FALSE)
}
\arguments{
\item{version}{Version number.  The default will load the most
recent version on your computer or the most recent version known
to the package if you have never downloaded the data before.
With \code{fw_del}, specifying \code{version=NULL} will
delete \emph{all} data sets.}

\item{path}{Path to store the data at.  If not given,
\code{datastorr} will use \code{rappdirs} to find the best place
to put persistent application data on your system.  You can
delete the persistent data at any time by running
\code{fw_del(NULL)} (or \code{fw_del(NULL, path)} if you
use a different path).}

\item{biomass}{Logical indicating whether or not you want the BWG
biomass dataset found at \url{https://github.com/SrivastavaLab/bwgbiomass}}
}
\description{
Download CESAB Functionalwebs Data
}
