% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data.R
\name{fw_version_current}
\alias{fw_version_current}
\title{Find current version}
\usage{
fw_version_current(local = TRUE, path = NULL)
}
\arguments{
\item{local}{Logical indicating if local or github versions should
be polled.}
}

