% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_data.R
\name{fw_versions}
\alias{fw_versions}
\title{Get available versions of data}
\usage{
fw_versions(local = TRUE, path = NULL)
}
\arguments{
\item{local}{Logical indicating if local or github versions should
be polled.  With any luck, \code{local=FALSE} is a superset of
\code{local=TRUE}.  For \code{fw_version_current}, if
\code{TRUE}, but there are no local versions, then we do check
for the most recent github version.}
}

